(Optionally) Don't Create Logs
Everyone loves logs, right? Well, maybe there's a time when creating a log
is unnecessary (for example, if the output of the script is being piped to
a dedicated log file). To support such cases, a "nologs" option has been
added to the script which will stop an external log file being written or
appended to.